+++ src/common/TreeNodeInterface.php
@@ -3,6 +3,7 @@ namespace VovanVE\parser\common;
 
 interface TreeNodeInterface
 {
+    /** @deprecated */
     const DUMP_INDENT = '    ';
 
     /**
@@ -11,10 +12,24 @@ interface TreeNodeInterface
     public function getNodeName();
 
     /**
+     * @return string|null
+     * @since 1.3.0
+     */
+    public function getNodeTag();
+
+    /**
      * @return integer
      * @since 1.1.0
      */
     public function getChildrenCount();
+
+    /**
+     * @param int $index Zero based
+     * @return TreeNodeInterface
+     * @throws \OutOfRangeException No child node with such index
+     * @since 1.3.0
+     */
+    public function getChild($index);
 
     /**
      * @return TreeNodeInterface[]
@@ -35,4 +50,16 @@ interface TreeNodeInterface
      * @return string
      */
     public function dumpAsString($indent = '', $last = true);
+
+    /**
+     * @param mixed $value
+     * @since 1.3.0
+     */
+    public function make($value);
+
+    /**
+     * @return mixed
+     * @since 1.3.0
+     */
+    public function made();
 }