+++ src/tree/NonTerminal.php
@@ -6,10 +6,38 @@ use VovanVE\parser\common\TreeNodeInterface;
 
 class NonTerminal extends BaseObject implements TreeNodeInterface
 {
-    /** @var string */
+    /**
+     * @var string
+     * @deprecated Don't use outside directly - use getter
+     */
     public $name;
-    /** @var TreeNodeInterface[] */
+    /** @var string|null */
+    private $tag;
+    /**
+     * @var TreeNodeInterface[]
+     * @deprecated Don't use outside directly - use getter
+     */
     public $children;
+    /** @var mixed */
+    private $made;
+
+    /**
+     * @param string $name
+     * @param TreeNodeInterface[] $children
+     * @param string|null $tag
+     * @since 1.3.0
+     */
+    public function __construct($name, $children, $tag = null)
+    {
+        $this->name = $name;
+        $this->children = $children;
+        if (null !== $tag) {
+            $tag = (string)$tag;
+            if ('' !== $tag) {
+                $this->tag = $tag;
+            }
+        }
+    }
 
     /**
      * @inheritdoc
@@ -21,11 +49,32 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
 
     /**
      * @inheritdoc
+     * @since 1.3.0
+     */
+    public function getNodeTag()
+    {
+        return $this->tag;
+    }
+
+    /**
+     * @inheritdoc
      * @since 1.1.0
      */
     public function getChildrenCount()
     {
         return count($this->children);
+    }
+
+    /**
+     * @inheritdoc
+     * @since 1.3.0
+     */
+    public function getChild($index)
+    {
+        if ($index >= 0 && $index < count($this->children)) {
+            return $this->children[$index];
+        }
+        throw new \OutOfBoundsException('No children');
     }
 
     /**
@@ -43,8 +92,7 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
      */
     public function areChildrenMatch($nodeNames)
     {
-        $count = $this->getChildrenCount();
-        if (count($nodeNames) !== $count) {
+        if (count($nodeNames) !== $this->getChildrenCount()) {
             return false;
         }
 
@@ -65,12 +113,36 @@ class NonTerminal extends BaseObject implements TreeNodeInterface
      */
     public function dumpAsString($indent = '', $last = true)
     {
-        $out = $indent . ' `- ' . $this->name . PHP_EOL;
+        $out = $indent . ' `- ' . $this->name;
+
+        if (null !== $this->tag) {
+            $out .= '(' . $this->tag . ')';
+        }
+
+        $out .= PHP_EOL;
         $sub_indent = $indent . ($last ? '    ' : ' |  ');
         $last_i = count($this->children) - 1;
         foreach ($this->children as $i => $child) {
             $out .= $child->dumpAsString($sub_indent, $i === $last_i);
         }
         return $out;
+    }
+
+    /**
+     * @inheritdoc
+     * @since 1.3.0
+     */
+    public function make($value)
+    {
+        $this->made = $value;
+    }
+
+    /**
+     * @inheritdoc
+     * @since 1.3.0
+     */
+    public function made()
+    {
+        return $this->made;
     }
 }